bugfix in torch backend, complex type casting
The syntax for type-casting in numpy and torch is not the same.
np.complex128() can take arguments, but if non passed return empty array
Will not return dtype!
If one passed it will be used as real part.
torch.complex() always needs to get real and complex part.
So if only real exist imag has to be initateded as zerros.